Restructured graphinfo
